Implement mutator to hold onto a stage.
A database implementation will need to hold onto a stage in order to
prevent it from being merged until commit. Also added a notification
for an iterator of the held stage to indicate that it has read a new set
of records from the stage. Comments in code explain why this is
necessary, see change set.

Closes #81.