Mark conflicts.
Provide a set of valid versions in order to mark conflicts. If you
update a record that has a version not in the set, the mutator is marked
as conflicted. It doesn't raise an exception, it proceeds with the
operation. You can decide how you want to report conflicts if at all.

If you provide no set of versions, no conflict checking is performed.

Closes #100.
See #62.