Upgrade `skip` to `0.2.5`.
The new Skip will search across pages for additional matches to the
filter function. This is used to find all the versions of a particular
record but if the user is using partial functions and wants to give us a
partial key and a filter for it, we can oblige. We'll be using this in
Memento to implement `map` against indices. Indices may or may not be
unique so for a given sought key we can return multiple results.

Closes #182.