User specified `group` function to `map`.
Pass a user defined `group` function for boundary detection through to
Skip. As noted in the previous log entry, with a link to the release
with more details the how and why, the new Skip allows for a user
specified `group` function.

Out unit test does not do any real grouping, just checks that we're
passing the key correctly. Memento's unit tests will ensure that it
works. Our unit tests does ensure that we're peeling off the correct
user key.

Closes #184.