Unwrap sought key in map of staging.
When we use the Skip module to create a map of staging to merge into a
common map we wrap the key in an array because the staging key is a
compound key of the users key plus version information. To create a
consistent sought structure to return the user we need to unwrap the
sought key returned from Skip for the staging invocations.

Closes #188.