Upgrade `designate` to `2.0.0-alpha.10`.
Updated to use new Skip module that groups matched items in an array for
each entry in the set to map. This change required updates to
Homogenize as well, so this upgrade will fix the previous broken builds.
The change logs for Skip release v0.2.7, Homogenize release
v2.0.0-alpha.11 and Designate release v2.0.0-alpha.10 contain details
about the changes.

https://github.com/bigeasy/designate/releases/tag/v2.0.0-alpha.10
https://github.com/bigeasy/homogenize/releases/tag/v2.0.0-alpha.11
https://github.com/bigeasy/skip/releases/tag/v0.2.7

Closes #193.