Update for iterator refactor.
Iterators now have a type property that is one of `mvcc.FORWARD`,
`mvcc.REVERSE` or `mvcc.MAP`. This means that iterator constructors that
decorate an iterator and need to create an iterator based on the
direction of the iterator no longer provide a `.forward()` and
`.reverse()` constructor function. Instead they inspect the type
property.

Advance and Riffle which are not decorators and actually iterate an
array or b-tree respectively now have a `reverse` named argument to
create a reverse iterator. Advance additionally has a `map` property if
the array contains a map structured set of entries. Skip always creates
a `mvcc.MAP` type.

Reciprocate 0.0.2 has an error-first callback function that will run the
trampoline for use in Locket implements the LevelDB interface which is
all error-first callbacks.

https://github.com/bigeasy/strata/releases/tag/v2.0.0-alpha.64
https://github.com/bigeasy/designate/releases/tag/v2.0.0-alpha.12
https://github.com/bigeasy/dilute/releases/tag/v2.0.0-alpha.6
https://github.com/bigeasy/homogenize/releases/tag/v2.0.0-alpha.12
https://github.com/bigeasy/reciprocate/releases/tag/v0.0.2
https://github.com/bigeasy/riffle/releases/tag/v2.0.0-alpha.43
https://github.com/bigeasy/riffle/releases/tag/v2.0.0-alpha.43
https://github.com/bigeasy/skip/releases/tag/v0.2.9
https://github.com/bigeasy/skip/releases/tag/v0.2.9
https://github.com/bigeasy/splice/releases/tag/v3.0.0-alpha.29
https://github.com/bigeasy/twiddle/releases/tag/v2.0.0-alpha.7

Closes #195.