Create `open` function.
Provide an static asynchrnous `open` function instead of a `ready`
property.

You'll notice that I'm awaiting `rejected` in destructible even though
it is returned from `createAmalgamator` in the test scripts. Learned
about the macrotask versus microtask queue and determined that the
return will be as a result of a `finally` run from the microtask queue
so we'll await `rejected` before the microtask queue completes. The
check for unhandled rejection occurs only when the microtask queue
completes.

Closes #197.