Correct reverse inclusive/exclusive.
Upstream, in `rifffle`, we did not correctly implement the less than or
equal to of reverse. It would start at the first entry greater the
default of a Strata search. Now that this is fixed, we see that we've
incorrectly our keys for inclusive reverse iteration. Or, rather, we did
not update our key selection when we changed the sort order of the
version from ascending to descending.

Closes #240.